requirements.txt
ollama
chromadb
fastapi
uvicorn
python-multipart
pymupdf
python-pptx
python-docx
pillow
langchain-text-splitters
pydantic
requests
pytest
junit-xml
openai
python-dotenv
opencv-python
numpy
pandas
oracledb
elasticsearch
# --- Embedding (local fallback for ollama provider) ---
sentence-transformers
# --- Agentic layer ---
langgraph
langchain-core
langchain-openai
langchain-google-genai
langchain-ollama